requirements.txt
fastapi>=0.118.0
inngest>=0.5.9
llama-index-core>=0.14.3
llama-index-readers-file>=0.5.4
openai>=1.109.1
python-dotenv>=1.1.1
qdrant-client>=1.15.1
requests>=2.31.0
uvicorn>=0.37.0

